requirements.txt
flask
line-bot-sdk
firebase-admin
google-generativeai
gunicorn
pandas
newsapi-python